requirements.txt
openai
SpeechRecognition
pygame
requests
python-dotenv
